provideSyncRefreshControl
